[Fix] build 에러 수정
